pyproject.toml
[project]
name = "finetune-on-aws"
version = "0.1.0"
description = "LLM Fine-tuning on AWS using SageMaker"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.37.38",
    "requests>=2.31.0",
    "flask>=2.3.0",
    "jupyter",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = []
